# openapi.json
{
  "components": {
    "schemas": {
      "CaseOverErrorBody": {
        "additionalProperties": false,
        "description": "409 body returned when the Case is already terminal.",
        "properties": {
          "detail": {
            "default": "case_over",
            "title": "Detail",
            "type": "string"
          }
        },
        "title": "CaseOverErrorBody",
        "type": "object"
      },
      "CaseResponse": {
        "additionalProperties": false,
        "description": "Blind in-progress response for GET /cases/{id} and POST /cases/{id}/move.",
        "properties": {
          "clock": {
            "title": "Clock",
            "type": "string"
          },
          "day": {
            "title": "Day",
            "type": "integer"
          },
          "detective_location": {
            "title": "Detective Location",
            "type": "string"
          },
          "hour": {
            "title": "Hour",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "day",
          "hour",
          "clock",
          "detective_location",
          "status"
        ],
        "title": "CaseResponse",
        "type": "object"
      },
      "CreateCaseRequest": {
        "additionalProperties": false,
        "description": "Body for POST /cases.",
        "properties": {
          "scenario": {
            "title": "Scenario",
            "type": "string"
          },
          "seed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          }
        },
        "required": [
          "scenario"
        ],
        "title": "CreateCaseRequest",
        "type": "object"
      },
      "CreateCaseResponse": {
        "additionalProperties": false,
        "description": "Response for POST /cases.",
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "detective_location": {
            "title": "Detective Location",
            "type": "string"
          },
          "locations": {
            "items": {
              "$ref": "#/components/schemas/LocationOut"
            },
            "title": "Locations",
            "type": "array"
          }
        },
        "required": [
          "case_id",
          "detective_location",
          "locations"
        ],
        "title": "CreateCaseResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "additionalProperties": false,
        "description": "Liveness payload polled by the TUI launcher before it connects (ADR-0007).",
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "IllegalMoveErrorBody": {
        "additionalProperties": false,
        "description": "400 body returned for non-adjacent / self / unknown move targets.",
        "properties": {
          "detail": {
            "default": "illegal_move",
            "title": "Detail",
            "type": "string"
          }
        },
        "title": "IllegalMoveErrorBody",
        "type": "object"
      },
      "LocationOut": {
        "additionalProperties": false,
        "description": "A named Location as returned to clients (Escape Location excluded).",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "neighbors": {
            "items": {
              "type": "string"
            },
            "title": "Neighbors",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "LocationOut",
        "type": "object"
      },
      "MoveRequest": {
        "additionalProperties": false,
        "description": "Body for POST /cases/{id}/move.",
        "properties": {
          "target": {
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "target"
        ],
        "title": "MoveRequest",
        "type": "object"
      },
      "TerminalCaseResponse": {
        "additionalProperties": false,
        "description": "Terminal response for POST /cases/{id}/move when the case is over.",
        "properties": {
          "clock": {
            "title": "Clock",
            "type": "string"
          },
          "day": {
            "title": "Day",
            "type": "integer"
          },
          "detective_location": {
            "title": "Detective Location",
            "type": "string"
          },
          "fugitive_route": {
            "items": {
              "type": "string"
            },
            "title": "Fugitive Route",
            "type": "array"
          },
          "hour": {
            "title": "Hour",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "day",
          "hour",
          "clock",
          "detective_location",
          "status",
          "fugitive_route"
        ],
        "title": "TerminalCaseResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Cawmen Sandaigo",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/cases": {
      "post": {
        "description": "Start a new Case: seed the fugitive Route and return the opening state.",
        "operationId": "create_case_cases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCaseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Case"
      }
    },
    "/cases/{case_id}": {
      "get": {
        "description": "Return the blind in-progress state for `case_id`.",
        "operationId": "get_case_cases__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "accept-language",
            "required": false,
            "schema": {
              "default": "en",
              "title": "Accept-Language",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Case"
      }
    },
    "/cases/{case_id}/move": {
      "post": {
        "description": "Apply a detective Move to `case_id`, returning the new or terminal state.",
        "operationId": "move_case_cases__case_id__move_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "accept-language",
            "required": false,
            "schema": {
              "default": "en",
              "title": "Accept-Language",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CaseResponse"
                    },
                    {
                      "$ref": "#/components/schemas/TerminalCaseResponse"
                    }
                  ],
                  "title": "Response Move Case Cases  Case Id  Move Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IllegalMoveErrorBody"
                }
              }
            },
            "description": "Illegal move target"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseOverErrorBody"
                }
              }
            },
            "description": "Case already finished"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Move Case"
      }
    },
    "/health": {
      "get": {
        "description": "Report backend liveness.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    }
  }
}